separate config store files
