--- src/config.js
@@ -1,1 +1,0 @@
-export const apiUrl = '/api'
