+++ src/constants/config.js
@@ -1,0 +1,4 @@
+export const API_URL = '/api'
+export const CONFIG_REQUEST = 'CONFIG_REQUEST'
+export const CONFIG_SUCCESS = 'CONFIG_SUCCESS'
+export const CONFIG_FAILURE = 'CONFIG_FAILURE'
